temperature_sensor/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--extra-index-url https://www.piwheels.org/simple

django==2.2.4
pytz==2019.2              # via django
sqlparse==0.3.0           # via django
django-jchart
